Make declaration & definition of qHash for DownloadTask consistant

+++ src/qx-net.h
@@ -16,7 +16,7 @@ struct DownloadTask
     QFile* dest;
 
     friend bool operator== (const DownloadTask& lhs, const DownloadTask& rhs) noexcept;
-    friend size_t qHash(const DownloadTask& key, size_t seed) noexcept;
+    friend uint qHash(const DownloadTask& key, uint seed) noexcept;
 };
 
 //-Classes------------------------------------------------------------------------------------------------------------
